#11793 Delete User Dependency before delete the User https://github.com/metasfresh/metasfresh/issues/11793
