+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/security/IRoleDAO.java
@@ -49,4 +49,6 @@ public interface IRoleDAO extends ISingletonService
 	RoleId retrieveFirstRoleIdForUserId(UserId adUserId);
 
 	void createUserRoleAssignmentIfMissing(UserId adUserId, RoleId adRoleId);
+
+	void deleteUserRolesByUserId(final UserId userId);
 }
