+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/user/api/impl/UserDAO.java
@@ -18,15 +18,7 @@ import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.service.ClientId;
 import org.adempiere.util.proxy.Cached;
-import org.compiere.model.I_AD_PInstance;
-import org.compiere.model.I_AD_Preference;
 import org.compiere.model.I_AD_User;
-import org.compiere.model.I_AD_UserMail;
-import org.compiere.model.I_AD_UserQuery;
-import org.compiere.model.I_AD_User_OrgAccess;
-import org.compiere.model.I_AD_User_Roles;
-import org.compiere.model.I_AD_User_Substitute;
-import org.compiere.model.I_C_OrgAssignment;
 import org.compiere.util.Env;
 import org.slf4j.Logger;
 
@@ -326,48 +318,4 @@ public class UserDAO implements IUserDAO
 		return OrgMappingId.optionalOfRepoId(sourceUserRecord.getAD_Org_Mapping_ID());
 	}
 
-	@Override
-	public void deleteUserSubstituteByUserId(final UserId userId)
-	{
-		queryBL.createQueryBuilder(I_AD_User_Substitute.class)
-				.addEqualsFilter(I_AD_User_Substitute.COLUMNNAME_AD_User_ID, userId)
-				.create()
-				.delete();
-	}
-
-	@Override
-	public void deleteUserMailByUserId(final UserId userId)
-	{
-		queryBL.createQueryBuilder(I_AD_UserMail.class)
-				.addEqualsFilter(I_AD_UserMail.COLUMNNAME_AD_User_ID, userId)
-				.create()
-				.delete();
-	}
-
-	@Override
-	public void deleteUserQueryByUserId(final UserId userId)
-	{
-		queryBL.createQueryBuilder(I_AD_UserQuery.class)
-				.addEqualsFilter(I_AD_UserQuery.COLUMNNAME_AD_User_ID, userId)
-				.create()
-				.delete();
-	}
-
-	@Override
-	public void deleteUserRolesByUserId(final UserId userId)
-	{
-		queryBL.createQueryBuilder(I_AD_User_Roles.class)
-				.addEqualsFilter(I_AD_User_Roles.COLUMNNAME_AD_User_ID, userId)
-				.create()
-				.delete();
-	}
-
-	@Override
-	public void deleteUserOrgAssignmentByUserId(final UserId userId)
-	{
-		queryBL.createQueryBuilder(I_C_OrgAssignment.class)
-				.addEqualsFilter(I_C_OrgAssignment.COLUMNNAME_AD_User_ID, userId)
-				.create()
-				.delete();
-	}
 }
